- modularized the loader functionality - had A LOT of trouble figuring out how to link the loader.o; Basically, was passing the arguments as such to the linker:
clang -L. -lbpf... loader.o main.o -o main

when it should be

clang loader.o main.o -L. -lbpf -o main

Because the linker resolves from left to right, noting undefined symbols along the way. The file that needs the symbol comes
before the one that resolves it.

Since loader.o uses a bunch of libbpf functions, by placing it after libbpf, the linker saw libbf first, which is because it's a static library,
is self-contained - no problems there. Then it saw loader.o with bunch of symbols that do not define themselves. The linker has nothing after it to
resolve them - which is also why they suggest including the library twice, at the beginning and end, so that the linked would be able to resolve it then the second time, after it sees loader.o's symbols. Basically, the linked has "short memory" =) btw this second approach is dirty, not going with it - I prefer to just include the object files before libbpf.

https://stackoverflow.com/questions/45135/why-does-the-order-in-which-libraries-are-linked-sometimes-cause-errors-in-gcc